solenoids
lso a minor change to firmware tho not sure how difficult to execute this. Have a copy of the firmware that has a different sol configuration. It will be same as a previous version where when cr1 is detected it activates sol2. When cr2 then sol 1. But when both cr1 and 2 are activated together neither activate.